requirements.txt
argparse>=1.4.0
shodan>=1.26.0
python-dotenv>=0.18.0
